refactor(workspace): split asset add workflows
