+++ .sampo/changesets/957-split-workspace-assets.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split workspace asset add workflows into focused pattern, binding-source, and editor-plugin runtime modules.
